refactor: transfer the project files to the src folder
