+++ index.js
@@ -1,8 +1,8 @@
 const { pipeline } = require('stream');
-const readStream = require('./read-stream');
-const writeStream = require('./write-stream');
-const transformArray = require('./transform-array');
-const { config } = require('./args');
+const readStream = require('./src/read-stream');
+const writeStream = require('./src/write-stream');
+const transformArray = require('./src/transform-array');
+const { config } = require('./src/args');
 
 pipeline(readStream, ...transformArray(config), writeStream, (err) => {
   if (err) {
