+++ src/read-stream.js
