+++ src/validation/config-validation.js
