+++ src/write-stream.js
